Informacion adicional en el pdf

+++ database/migrations/tenant/2018_06_17_000005_tenant_document_items_table.php
@@ -19,7 +19,7 @@ class TenantDocumentItemsTable extends Migration
             $table->unsignedInteger('document_id');
             $table->unsignedInteger('item_id');
             $table->json('item');
-            $table->integer('quantity');
+            $table->decimal('quantity', 12, 2);
             $table->decimal('unit_value', 12, 2);
 
             $table->string('affectation_igv_type_id');
